backend/requirements.txt
wheel==0.37.0
gunicorn==20.0.4
Django==3.2.9
psycopg2-binary==2.9.1
django-environ==0.8.1
djangorestframework==3.12.4
pillow==8.4.0
djoser==2.1.0
reportlab==3.6.2
python-environ==0.4.54
django-filter==21.1
django-extra-fields==3.0.2
